movedToSeperateClass
